add cache

+++ src/services/postgres/NotesService.js
@@ -6,11 +6,13 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 const { mapDBToModel } = require('../../utils');
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     // eslint-disable-next-line no-underscore-dangle
     this._pool = new Pool();
     // eslint-disable-next-line no-underscore-dangle
     this._collaborationService = collaborationService;
+    // eslint-disable-next-line no-underscore-dangle
+    this._cacheService = cacheService;
   }
 
   async addNote({
@@ -34,21 +36,37 @@ class NotesService {
     if (!result.rows[0].id) {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
+    // Menghapus cache, karena ada perubahan/penambahan data notes
+    // eslint-disable-next-line no-underscore-dangle
+    await this._cacheService.delete(`notes:${owner}`);
 
     return result.rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-      LEFT JOIN collaborations ON collaborations.note_id = notes.id
-      WHERE notes.owner = $1 OR collaborations.user_id = $1
-      GROUP BY notes.id`,
-      values: [owner],
-    };
-    // eslint-disable-next-line no-underscore-dangle
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModel);
+    try {
+      // mendapatkan catatan dari cache
+      // eslint-disable-next-line no-underscore-dangle
+      const result = await this._cacheService.get(`notes:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+        LEFT JOIN collaborations ON collaborations.note_id = notes.id
+        WHERE notes.owner = $1 OR collaborations.user_id = $1
+        GROUP BY notes.id`,
+        values: [owner],
+      };
+      // eslint-disable-next-line no-underscore-dangle
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      // catatan akan disimpan pada cache sebelum fungsi getNotes dikembalikan
+      // eslint-disable-next-line no-underscore-dangle
+      await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -70,7 +88,7 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id',
+      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
     // eslint-disable-next-line no-underscore-dangle
@@ -79,11 +97,15 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+    // eslint-disable-next-line no-underscore-dangle
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM NOTES WHERE id=$1 RETURNING id',
+      text: 'DELETE FROM NOTES WHERE id=$1 RETURNING id, owner',
       values: [id],
     };
     // eslint-disable-next-line no-underscore-dangle
@@ -92,6 +114,10 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+    // eslint-disable-next-line no-underscore-dangle
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async verifyNoteOwner(id, owner) {
